sdk/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openshift-ai-sdk"
version = "0.1.0"
description = "OpenShift AI SDK for managing training jobs, including MPIJob support"
readme = "README.md"
authors = [
    {name = "OpenShift AI Team", email = "openshift-ai@redhat.com"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "kubernetes>=24.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=2.0.0",
    "pytest-mock>=3.0.0",
]
dev = [
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/openshift-ai/mpijob"
"Bug Tracker" = "https://github.com/openshift-ai/mpijob/issues"

[tool.setuptools]
packages = ["odh"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true